+++ services/assessment/models/models.go
@@ -43,7 +43,7 @@ type AssessmentAttempt struct {
 	Status           string             `json:"string" pg:",notnull"`
 	StartedAt        *time.Time         `json:"started_at,omitempty"`
 	CompletedAt      *time.Time         `json:"completed_at,omitempty"`
-	Score            uint32             `json:"score,omitempty" pg:",use_zero"`
+	Score            int64              `json:"score,omitempty" pg:",use_zero"`
 	Assessment       *Assessment        `json:"assessment" pg:"rel:has-one"`
 	Questions        []*Question        `json:"questions" pg:",many2many:attempts_questions,fk:attempt_id,join_fk:question_id"`
 	QuestionAttempts []*AttemptQuestion `json:"question_attempts" pg:"rel:has-many,join_fk:attempt_id"`
@@ -60,7 +60,7 @@ type Question struct {
 	MediaURL           string               `json:"image_url" pg:"media_url"`
 	Code               string               `json:"code"`
 	Options            []string             `json:"options" pg:",array"`
-	Answer             uint32               `json:"answer" pg:",use_zero"`
+	Answer             int64                `json:"answer" pg:",use_zero"`
 	Tags               []*Tag               `json:"tags" pg:"rel:has-many"`
 	Assessments        []*Assessment        `json:"assessments" pg:"many2many:assessments_questions"`
 	AssessmentAttempts []*AssessmentAttempt `json:"assessment_attempts" pg:",many2many:attempts_questions,fk:question_id,join_fk:attempt_id"`
@@ -75,15 +75,16 @@ type AttemptQuestion struct {
 	AttemptID   uint64     `json:"attempt_id" pg:"attempt_id,notnull"`
 	QuestionID  uint64     `json:"question_id" pg:"question_id,notnull"`
 	CandidateID uint64     `json:"candidate_id" pg:"candidate_id,notnull"`
-	Selection   uint32     `json:"selection" pg:",use_zero"`
+	Selection   int64      `json:"selection" pg:",use_zero"`
 	Text        string     `json:"text"`
 	CMMode      string     `json:"cm_mode"`
-	Score       uint32     `json:"score,omitempty" pg:",use_zero"`
+	Score       int64      `json:"score,omitempty" pg:",use_zero"`
 	TimeTaken   uint64     `json:"time_taken,omitempty"`
 	CreatedAt   *time.Time `json:"created_at"`
 	UpdatedAt   *time.Time `json:"updated_at"`
 }
 
+// BeforeInsert handles the event before an AttemptQuestion is inserted into the DB
 func (m *AttemptQuestion) BeforeInsert(ctx context.Context) (context.Context, error) {
 	now := time.Now()
 	m.CreatedAt = &now
@@ -91,6 +92,7 @@ func (m *AttemptQuestion) BeforeInsert(ctx context.Context) (context.Context, er
 	return ctx, nil
 }
 
+// BeforeUpdate handles the event before an AttemptQuestion is updated in the DB
 func (m *AttemptQuestion) BeforeUpdate(ctx context.Context) (context.Context, error) {
 	now := time.Now()
 	m.UpdatedAt = &now
